pyproject.toml
[project]
name = "PyKMCMC"
version = "0.1.0"
description = ""
authors = [
    {name = "Nathaniel D. Hoffman", email = "dene@cmu.edu"},
]
dependencies = [
    "numpy>=1.25.2",
    "uproot>=5.0.11",
    "pandas>=2.1.0",
    "jax[cpu]>=0.4.14",
    "zeus-mcmc>=2.5.4",
    "rich>=13.5.2",
    "emcee>=3.1.4",
    "docopt-ng>=0.9.0",
    "schwimmbad>=0.3.2",
    "h5py>=3.9.0",
    "codetiming>=1.4.0",
    "corner>=2.2.2",
    "iminuit>=2.24.0",
    "watchdog>=3.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
kmcmc = {call = 'pykmcmc.mcmc:run_mcmc'}
kplot = {call = 'pykmcmc.plotter:plotter'}
kmonitor = {call = 'pykmcmc.monitor:monitor'}
kzeus = {call = 'pykmcmc.zeus:run_zeus'}
kzplot = {call = 'pykmcmc.zplot:zplot'}
kfit = {call = 'pykmcmc.fit:run_fit'}
kwatch = {call = 'pykmcmc.live:watch'}

[tool.pyright]
extraPaths = ["__pypackages__/3.11/lib/"]
